feat(ingest/BigQuery): refractor+parallise dataset metadata extraction
